Ryddet i klagebehandling-controllerne. Tenker at path uttrykker noe om hvilken ressurs kallene agerer på, at at kall rundt samme ressurs bør samles i samme controller. La til en generell KlagebehandlingController, mulig at noe der kan flyttes inn i spesifikke etterhvert.
