--- src/main/kotlin/no/nav/klage/oppgave/api/controller/KlagebehandlingEditableFieldsController.kt
@@ -1,64 +1,0 @@
-package no.nav.klage.oppgave.api.controller
-
-import io.swagger.annotations.Api
-import no.nav.klage.oppgave.api.mapper.KlagebehandlingMapper
-import no.nav.klage.oppgave.api.view.KlagebehandlingEditableFieldsInput
-import no.nav.klage.oppgave.api.view.KlagebehandlingEditedView
-import no.nav.klage.oppgave.config.SecurityConfiguration.Companion.ISSUER_AAD
-import no.nav.klage.oppgave.exceptions.BehandlingsidWrongFormatException
-import no.nav.klage.oppgave.repositories.InnloggetSaksbehandlerRepository
-import no.nav.klage.oppgave.service.KlagebehandlingEditableFieldsFacade
-import no.nav.klage.oppgave.util.getLogger
-import no.nav.security.token.support.core.api.ProtectedWithClaims
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PutMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RestController
-import java.util.*
-
-@RestController
-@Api(tags = ["kabal-api"])
-@ProtectedWithClaims(issuer = ISSUER_AAD)
-class KlagebehandlingEditableFieldsController(
-    private val innloggetSaksbehandlerRepository: InnloggetSaksbehandlerRepository,
-    private val klagebehandlingMapper: KlagebehandlingMapper,
-    private val editableFieldsFacade: KlagebehandlingEditableFieldsFacade,
-) {
-
-    companion object {
-        @Suppress("JAVA_CLASS_ON_COMPANION")
-        private val logger = getLogger(javaClass.enclosingClass)
-    }
-
-    @PutMapping("/klagebehandlinger/{id}/detaljer/editerbare")
-    fun putEditableFields(
-        @PathVariable("id") klagebehandlingId: String,
-        @RequestBody input: KlagebehandlingEditableFieldsInput
-    ): KlagebehandlingEditedView {
-        logMethodDetails("putEditableFields", klagebehandlingId)
-        return klagebehandlingMapper.mapKlagebehandlingToKlagebehandlingEditableFieldsView(
-            editableFieldsFacade.updateEditableFields(
-                klagebehandlingId.toUUIDOrException(),
-                input,
-                innloggetSaksbehandlerRepository.getInnloggetIdent()
-            )
-        )
-    }
-
-    private fun String.toUUIDOrException() =
-        try {
-            UUID.fromString(this)
-        } catch (e: Exception) {
-            logger.error("Input could not be parsed as an UUID", e)
-            throw BehandlingsidWrongFormatException("Input could not be parsed as an UUID")
-        }
-
-    private fun logMethodDetails(methodName: String, klagebehandlingId: String) {
-        logger.debug(
-            "{} is requested by ident {} for klagebehandlingId {}",
-            methodName,
-            innloggetSaksbehandlerRepository.getInnloggetIdent(),
-            klagebehandlingId
-        )
-    }
-}
